refactor(RandomNumberFactory): 세 개의 랜덤 숫자를 뽑고, RandomNumber 클래스에 저장하도록 변경
세 개의 랜덤 숫자를 뽑아 그대로 반환해줄 경우, List에 값을 더하거나 빼는 등의 변경이 가능한 것을 확인했습니다.
세 개의 랜덤 숫자는 '정답'이 될 숫자이기 때문에 변해서는 안된다고 판단하였고, 방법을 찾아본 결과 '일급 컬렉션'이라는 개념을 찾았습니다.
이를 적용하여, 할당된 값이 절대 변경되지 않도록 리팩토링 하였습니다.
