+++ src/main/java/baseball/randomnumber/RandomNumber.java
@@ -1,0 +1,17 @@
+package baseball.randomnumber;
+
+import java.util.Collections;
+import java.util.List;
+
+public class RandomNumber {
+
+    private final List<Integer> answer;
+
+    RandomNumber(List<Integer> randomNumber) {
+        this.answer = randomNumber;
+    }
+
+    public List<Integer> getAnswer() {
+        return Collections.unmodifiableList(answer);
+    }
+}
